segmentation-master/requirements.txt
absl-py==0.9.0
certifi==2018.8.24
decorator==4.4.2
grpcio==1.30.0
imagecodecs==2019.12.31
imageio==2.9.0
joblib==0.14.1
Markdown==2.6.11
networkx==2.4
numpy==1.16.4
olefile==0.46
opencv-python==4.3.0.36
pandas==0.25.3
Pillow==5.2.0
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
tensorboard==1.14.0
tensorboardX==2.1
tifffile==2019.7.26.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
Werkzeug==1.0.1
wincertstore==0.2
